Split Computer Use platform code
